Added frequency extrapolation tooling (smart object cloning)
It is importatant to easily choose which frequencies will be used for
calculations because the functions below return vectors with values at each
frequency for the calculation provided.  In many cases you will want to load a
Touchstone data file at all frequencies and then use C<$obj-E<gt>at($f_Hz)> to
reduce or extrapolate to a different frequency or set of freqencies.

See POD for more.

Also:
	Added interpolation (->at) tests
	Added save_snp_fh
	Upstream removed freq_min_hz-based frequency ranges, so use the new format.